Add project level audio analysis page
